help: add gitsubmodules to the list of guides
The guide "gitsubmodules" was added in e88d731 (submodules: overhaul
documentation, 2017-06-22), but it was not added to
command-list.txt when commit 9803dc1 (help: use command-list.txt
for the source of guides, 2018-05-20) taught "git help" to obtain the
guide list from this file.

Add it now, and capitalize the first word of the description of
gitsubmodules, as was done in 9803dc1 (help: use command-list.txt
for the source of guides, 2018-05-20) for the other guides.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>
Reviewed-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>